LPS-168348 knowledge-base-service: Copy from KBArticleIndexer doGetDocument, no logic changes
https://issues.liferay.com/browse/LPS-168348
