+++ modules/apps/knowledge-base/knowledge-base-service/src/main/java/com/liferay/knowledge/base/internal/search/KBArticleIndexer.java
@@ -15,13 +15,9 @@
 package com.liferay.knowledge.base.internal.search;
 
 import com.liferay.document.library.kernel.model.DLFileEntry;
-import com.liferay.knowledge.base.constants.KBFolderConstants;
 import com.liferay.knowledge.base.model.KBArticle;
-import com.liferay.knowledge.base.model.KBFolder;
 import com.liferay.knowledge.base.service.KBArticleLocalService;
-import com.liferay.knowledge.base.service.KBFolderLocalService;
 import com.liferay.knowledge.base.util.KnowledgeBaseUtil;
-import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.dao.orm.IndexableActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.Property;
@@ -47,8 +43,6 @@ import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.HtmlParser;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 
@@ -131,20 +125,6 @@ public class KBArticleIndexer extends BaseIndexer<KBArticle> {
 	protected Document doGetDocument(KBArticle kbArticle) throws Exception {
 		Document document = getBaseModelDocument(CLASS_NAME, kbArticle);
 
-		document.addText(
-			Field.CONTENT, _htmlParser.extractText(kbArticle.getContent()));
-		document.addText(Field.DESCRIPTION, kbArticle.getDescription());
-		document.addKeyword(Field.FOLDER_ID, kbArticle.getKbFolderId());
-		document.addText(Field.TITLE, kbArticle.getTitle());
-		document.addKeyword(
-			Field.TREE_PATH,
-			StringUtil.split(kbArticle.buildTreePath(), CharPool.SLASH));
-		document.addKeyword("folderNames", _getKBFolderNames(kbArticle));
-		document.addKeyword(
-			"parentMessageId", kbArticle.getParentResourcePrimKey());
-		document.addKeyword("titleKeyword", kbArticle.getTitle(), true);
-		document.addKeywordSortable("urlTitle", kbArticle.getUrlTitle());
-
 		return document;
 	}
 
@@ -216,25 +196,6 @@ public class KBArticleIndexer extends BaseIndexer<KBArticle> {
 	@Reference
 	protected KBArticleLocalService kbArticleLocalService;
 
-	@Reference
-	protected KBFolderLocalService kbFolderLocalService;
-
-	private String[] _getKBFolderNames(KBArticle kbArticle) throws Exception {
-		long kbFolderId = kbArticle.getKbFolderId();
-
-		Collection<String> kbFolderNames = new ArrayList<>();
-
-		while (kbFolderId != KBFolderConstants.DEFAULT_PARENT_FOLDER_ID) {
-			KBFolder kbFolder = kbFolderLocalService.getKBFolder(kbFolderId);
-
-			kbFolderNames.add(kbFolder.getName());
-
-			kbFolderId = kbFolder.getParentKBFolderId();
-		}
-
-		return kbFolderNames.toArray(new String[0]);
-	}
-
 	private void _reindexAttachments(KBArticle kbArticle) throws Exception {
 		Indexer<DLFileEntry> indexer = IndexerRegistryUtil.nullSafeGetIndexer(
 			DLFileEntry.class);
@@ -295,9 +256,6 @@ public class KBArticleIndexer extends BaseIndexer<KBArticle> {
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		KBArticleIndexer.class);
-
-	@Reference
-	private HtmlParser _htmlParser;
 
 	@Reference(
 		target = "(model.class.name=com.liferay.knowledge.base.model.KBArticle)"
